Split out build constrained record & replay tests
for runtimeconfigurator_test.go.
